Retrofit Express 4/5 dual scheme
Test the package against both Express major lines instead of pinning to
4.x. The package no longer has any `express` runtime dependency to gate
on, so dual coverage simply becomes a matter of how the test suite
imports its app factory.

Dependencies:
- Drop `express` from `dependencies` (no longer required at runtime).
- Bump `supertest` to ^7 in `dependencies` (the prior ^6.3.1 was
  pinning the project to an older line for no functional reason).
- Replace the single `@types/express` / `@types/supertest` devDeps with
  the dual-alias scheme already used by the rest of the family:
  `express4` / `express5` via `npm:express@^4` / `npm:express@^5`, and
  the matching `@types/express4` / `@types/express5` aliases. Bump
  `@types/supertest` to ^7 to match the runtime bump.
- Keep a thin `@types/express: ^4` so that `import type { ... } from
  "express"` inside lib/ still resolves without paths mapping.

Test layout:
- Move the per-topic logic out of `test/<topic>.test.ts` into
  `test/lib/<topic>.ts`, each exporting a `run<Topic>Tests(label,
  express)` runner.
- Add `test/express4.test.ts` and `test/express5.test.ts` that import
  the express factory under the corresponding alias and call every
  runner.
- `test/lib/util.ts` mirrors the family convention with an
  `ExpressFactory = any` shape that satisfies both 4.x and 5.x app
  builders.

Lib adjustments are only those required by the supertest@7 type
surface: `_agent` is now `supertest.Agent` (the renamed
`SuperTest<any>`), and the composed handler is typed as the 2-arg
`RequestListener` shape that supertest hands to `http.createServer`.

Test counts go from 18 (one Express version) to 36 (both versions),
all passing.
