+++ package.json
@@ -10,16 +10,19 @@
     "Yusuke Kawasaki <u-suke@kawa.net>"
   ],
   "dependencies": {
-    "express": "^4.18.2",
     "express-intercept": "^1.2.2",
-    "supertest": "^6.3.1"
+    "supertest": "^7"
   },
   "devDependencies": {
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@rollup/plugin-sucrase": "^5.1.0",
-    "@types/express": "^4.17.17",
+    "@types/express": "^4",
+    "@types/express4": "npm:@types/express@^4",
+    "@types/express5": "npm:@types/express@^5",
     "@types/node": "^24.12.2",
-    "@types/supertest": "^2.0.16",
+    "@types/supertest": "^7",
+    "express4": "npm:express@^4",
+    "express5": "npm:express@^5",
     "rollup": "^4.60.2",
     "typescript": "^5.9.3"
   },
