--- test/failure.test.ts
@@ -1,40 +1,0 @@
-import {strict as assert} from "node:assert";
-import {describe, it} from "node:test";
-import {fileURLToPath} from "node:url";
-import express from "express";
-import type {RequestHandler} from "express";
-
-import {mwsupertest} from "../lib/middleware-supertest.ts";
-import type {MWSuperTest} from "../types/middleware-supertest.d.ts";
-
-const TITLE = fileURLToPath(import.meta.url).split("/").pop();
-
-/**
- * This is a test to test mwsupertest to capture assertion thrown during mwsupertest.
- */
-
-describe(TITLE, () => {
-    const handler: RequestHandler = (req, res) => res.send("SUCCESS");
-    const prefix = "something wrong with ";
-    const app = express().use(handler);
-
-    test("getString", mwsupertest(app).getString(async () => assert.ok(false, prefix + "getString")));
-
-    test("getBuffer", mwsupertest(app).getString(async () => assert.ok(false, prefix + "getBuffer")));
-
-    test("getResponse", mwsupertest(app).getString(async () => assert.ok(false, prefix + "getResponse")));
-
-    test("getRequest", mwsupertest(app).getString(async () => assert.ok(false, prefix + "getRequest")));
-
-    function test(title: string, testApp: MWSuperTest) {
-        it(title, async () => {
-            let error: Error;
-            try {
-                await testApp.get("/").expect(500);
-            } catch (e) {
-                error = e;
-            }
-            assert.ok(String(error).indexOf(prefix + title) > -1, String(error));
-        });
-    }
-});
